added custom backend thread and commandline/configfile configuration parameters to communicate with the TWS application using the TWS API.
The code is still very rudimentary and has a pthread usage bugs, but the /tws/hello URL will deliver the timestamp as reported by IB through the TWS API

The backend thread will automatically connect to the TWS application and when this fails or when the connection is broken due to comm errors or other reasons (e.g. TWS app termination) the backend thread will continue to try to connect with a configurable interval -- this means that we do not have to perform any manual activity when the TWS connection is lost for whatever reason.